pyproject.toml
[tool.poetry]
name = "html2reader"
version = "0.1.0"
description = "Send web pages to reader in fb2"
authors = ["Apollon76 <devcpy@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.0"
dropbox = "^11.29.0"
html2text = "^2020.1.16"
lxml = "^4.8.0"
pocket-api = "^0.1.5"
pony = "^0.7.16"

[tool.poetry.dev-dependencies]
black = ">19.3b0"
flake8-awesome = "<1.3"
mypy = "^0.812"
pytest = "^6.2.4"
pytest-cov = "^2.10.1"
isort = "^5.10.1"

[tool.black]
target-version = ['py39']
line-length = 120

[build-system]
requires = ["poetry>=1.0"]
